backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
pydantic-settings==2.6.1
openai==1.54.0
httpx==0.27.2
chromadb==0.5.4
sentence-transformers==3.1.1
pypdf==5.1.0
